修改变量定义位置，内置

+++ Project2/my_class_fun.h
@@ -24,12 +24,12 @@ struct ModelMesh
 namespace psykronix
 {
 
-    static constexpr size_t N_10E_3 = 1e3;
-    static constexpr size_t N_10E_4 = 1e4;
-    static constexpr size_t N_10E_5 = 1e5;
-    static constexpr size_t N_10E_6 = 1e6;
-    static constexpr size_t N_10E_7 = 1e7;
-    static constexpr size_t N_10E_8 = 1e8;
+    static constexpr size_t N_10E_3 = (size_t)1e3;
+    static constexpr size_t N_10E_4 = (size_t)1e4;
+    static constexpr size_t N_10E_5 = (size_t)1e5;
+    static constexpr size_t N_10E_6 = (size_t)1e6;
+    static constexpr size_t N_10E_7 = (size_t)1e7;
+    static constexpr size_t N_10E_8 = (size_t)1e8;
 
     class Vertex
     {
@@ -71,7 +71,7 @@ namespace psykronix
         }
         operator Eigen::Vector3f() const
         {
-            return Eigen::Vector3f(x, y, z);
+            return Eigen::Vector3f((float)x, (float)y, (float)z);
         }
         //auto operator[](int i)
         //{
